+++ tests/variables.test.mjs
@@ -37,6 +37,7 @@ import {
     isset,
     unserialize,
     var_export,
+    to_boolean,
 } from './../src/variables.mjs';
 
 test('variables', () => {
@@ -59,6 +60,7 @@ test('variables', () => {
     expect(is_callable).not.toBeUndefined();
     expect(to_array).not.toBeUndefined();
     expect(to_bool).not.toBeUndefined();
+    expect(to_boolean).not.toBeUndefined();
     expect(floatval).not.toBeUndefined();
     expect(doubleval).not.toBeUndefined();
     expect(empty).not.toBeUndefined();